+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 1.3.0
 =============
 
 - Fixed some special NVMe capacity parsing issues (issue [#73](https://github.com/truenas/py-SMART/issues/73)). Thanks @wolfdate25
+- Other minnor fixes with @ulmitov's help
 - **Breaking changes**
     - **smartctl.all**:
         - Officially removed support for python 3.7. However, pyproject.toml still allows the 3.7 version. This will be removed in the next release.